requirements.txt
django==4.2.5
psycopg2_binary==2.9.7
djangorestframework==3.14.0
django-filter==23.2
Pillow==10.0.0
djangorestframework-simplejwt=5.3.0
django-extensions=3.2.3
drf-nested-routers=0.93.4
